requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
markdown2==2.4.12
python-slugify==8.0.4
azure-storage-blob==12.19.0
python-dotenv==1.0.0
gunicorn==21.2.0